app.js
In this repository you can see that the main character of the project is Frank Sinatra. The code that I wrote in this repository represents the backend side of the web page developement process. This code does multiple tasks. 

1) In first part of the code the user can get a random Frank SInatra's song out of his 20 most famous songs.

2) In second part of the code we move forward and create a multiple pages or routes. Here, after an interaction with the web page, customer can get an information about Frank Sinatra's date of birth, city where he was born and the list of his 4 wifes.

3) The third and the last part of the code responsible for log in part of the web page:

GET on /public. This action will print "Everybody can see this page"

GET on /protected. This action will be protected by a HTTP Basic access authentication and print "Welcome, authenticated client" if you are authorized with the login admin and password admin otherwise it will provide a 401 Not authorized.